+++ lib/types/highcharts_global_options.dart
@@ -32,9 +32,8 @@ export 'highcharts_global_button_theme_options.dart';
  * */
 
 /**
- * Global options that don't apply to each chart. These options, like
- * the `lang` options, must be set using the `Highcharts.setOptions`
- * method.
+ * Global options that don't apply to each chart. These options must be set
+ * using the `Highcharts.setOptions` method.
  * 
  * ```js
  * Highcharts.setOptions({