+++ lib/types/highcharts_series_last_price_options.dart
@@ -37,10 +37,13 @@ export 'highcharts_series_last_price_label_options.dart';
  */
 class HighchartsSeriesLastPriceOptions extends HighchartsOptionsBase {
   String? color;
+  String? dashStyle;
   bool? enabled;
   HighchartsSeriesLastPriceLabelOptions? label;
+  double? width;
 
-  HighchartsSeriesLastPriceOptions({this.color, this.enabled, this.label});
+  HighchartsSeriesLastPriceOptions(
+      {this.color, this.dashStyle, this.enabled, this.label, this.width});
 
   @override
   void toOptionsJSON(StringBuffer buffer) {
@@ -49,11 +52,17 @@ class HighchartsSeriesLastPriceOptions extends HighchartsOptionsBase {
     if (color != null) {
       buffer.writeAll(['"color":', jsonEncode(color), ','], '');
     }
+    if (dashStyle != null) {
+      buffer.writeAll(['"dashStyle":', jsonEncode(dashStyle), ','], '');
+    }
     if (enabled != null) {
       buffer.writeAll(['"enabled":', enabled, ','], '');
     }
     if (label != null) {
       buffer.writeAll(['"label":', label?.toJSON(), ','], '');
+    }
+    if (width != null) {
+      buffer.writeAll(['"width":', width, ','], '');
     }
   }
 }