+++ lib/types/highcharts_treegraph_series_data_labels_options.dart
@@ -64,6 +64,7 @@ class HighchartsTreegraphSeriesDataLabelsOptions extends HighchartsOptionsBase {
   HighchartsSeriesDataLabelsFilterOptions? filter;
   String? format;
   dynamic formatter;
+  bool? headers;
   bool? inside;
   HighchartsTreegraphSeriesDataLabelsLinkTextPathOptions? linkTextPath;
   String? nullFormat;
@@ -99,6 +100,7 @@ class HighchartsTreegraphSeriesDataLabelsOptions extends HighchartsOptionsBase {
       this.filter,
       this.format,
       this.formatter,
+      this.headers,
       this.inside,
       this.linkTextPath,
       this.nullFormat,
@@ -169,6 +171,9 @@ class HighchartsTreegraphSeriesDataLabelsOptions extends HighchartsOptionsBase {
     }
     if (formatter != null) {
       buffer.writeAll(['"formatter":', jsonEncode(formatter), ','], '');
+    }
+    if (headers != null) {
+      buffer.writeAll(['"headers":', headers, ','], '');
     }
     if (inside != null) {
       buffer.writeAll(['"inside":', inside, ','], '');